improve e2e
