+++ src/components/activity/ActivityFilter.tsx
@@ -8,10 +8,8 @@ import { cn, getLowestBits } from 'utils/helpers';
 import { ReactComponent as IconSearch } from 'assets/icons/search.svg';
 import { ReactComponent as IconPair } from 'assets/icons/token-pair.svg';
 import { TokensOverlap } from 'components/common/tokensOverlap';
-import {
-  DateRangePicker,
-  datePickerPresets,
-} from 'components/common/datePicker/DateRangePicker';
+import { DateRangePicker } from 'components/common/datePicker/DateRangePicker';
+import { datePickerPresets } from 'components/common/datePicker/utils';
 import { useActivity } from './ActivityProvider';
 
 interface DisplayID {
